Refactor: move color logic into FormFieldColor
