+++ src/components/settings/ColorPickerModal.jsx
@@ -1,9 +1,10 @@
 import PropTypes from "prop-types"
 import React, { useState } from "react"
 import { SketchPicker } from "react-color"
+
 import elementStyles from "styles/isomer-cms/Elements.module.scss"
 
-const ColorPicker = ({ value, onColorSelect, setRealTimeColor }) => {
+const ColorPickerModal = ({ value, onColorSelect, setRealTimeColor }) => {
   // initiate selectedColor as color passed in as props
   const [selectedColor, setSelectedColor] = useState(value)
   const handleColorSelect = (color, event) => {
@@ -37,9 +38,9 @@ const ColorPicker = ({ value, onColorSelect, setRealTimeColor }) => {
   )
 }
 
-export default ColorPicker
+export default ColorPickerModal
 
-ColorPicker.propTypes = {
+ColorPickerModal.propTypes = {
   value: PropTypes.string.isRequired,
   onColorSelect: PropTypes.func.isRequired,
   setRealTimeColor: PropTypes.func.isRequired,
